requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.1.4
numpy==1.26.3
prophet==1.1.5
transformers==4.36.2
torch==2.1.2
groq==0.4.1
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6
